+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 0.2-SNAPHSOT
 ## Added
 * `Database.getServer()`
+* `DocumentStub.contains(String)`